[SD-148] Split AVLp into modules
