+++ snowdrop-execution/src/Snowdrop/Execution/DbActions.hs
@@ -1,2 +1,1 @@
-
 {-# OPTIONS_GHC -F -pgmF autoexporter #-}
